requirements.txt
aiogram==3.4.1
python-gitlab==4.4.0
python-dotenv==1.0.0
aiohttp==3.9.1